students/k3343/Otroschenko_Valeria/Lr3/requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
